Updated git submodules
Project: docs stable/boron e5db1942cbf1880b94f757938df96b54736364e3

Updated git submodules

Project: odlparent master 3af246789fb39299a3acccb5e9bfb1da7856a06d

Bump maven-javadoc-plugin version from 2.10.3 to 2.10.4

Not just for fun, but because I've just figured, through trial and
error, that this fixes a weird problem seen in the failed builds of
https://git.opendaylight.org/gerrit/#/c/46206/ - the older 2.10.3 seems
to have some bug which causes it to choke on finding references between
(two different kinds of) generated code in JavaDoc gen. and the 2.10.4
upgrade magically fixes it (I've locally tested this).

Note that we seem to have our maven-javadoc-plugin version defined in
two places, which is a bit of a shame (and very confusing at first).  I
wanted to get this bump merged in ASAP, with minimal side effect, so in
this commit have just changed both places.  A follow-up commit which
unifies this into a single place for the next time around would be cool.

Change-Id: Ie2c35f4a7dea357eed5ce8b72c9e757dbdf85565
Signed-off-by: Michael Vorburger <vorburger@redhat.com>