Updated git submodules
Project: sfc master 984f23df2e09c607647d02a680610242da2d2caa

Remove Maven warnings for using variable in version

This commit fixes the warning for using a variable in the
version tag of the artifact.
The artifact should get the version from the parent pom.
Note that variables in version tag should not be allowed.

Change-Id: I2cafc4609e27c74ea0874b4a0c3bbf4fc1707cdd
Signed-off-by: Miguel Angel Munoz Gonzalez <miguel.angel.munoz.gonzalez@ericsson.com>