Updated git submodules
Project: netvirt master aaa192bf0bfcb105b319225d34114c81cb392237

Ethernet VPN Enhancement in BGP and FIB APIs

This review brings in API changes in FIB and BGP manager to support BGP
Ethernet VPN feature in OpenDaylight. This is the fifth in a series of
reviews with 'EVPN_RT5' as the topic.

This review brings in the following things :
1) FIB manager API changes to supply additional parameters for VPN to BGP,
when an L3VPN over VxLAN is created.
2) BGP manager API changes to confirm to thrift API changes and to supply
routes to FIB when BGP gets routes belonging to IP-VRF table of an EVPN.

Change-Id: I72efaa7f079b5aff31774167f10e867a3deea4bc
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>