Updated git submodules
Project: bgpcep stable/boron e07ac3381e49f1b90423bc11008a701df6da64ed

Do not wrap Guava as a bundle in features' definition

As it will refresh the bundle when the feature is loaded,
because it will certainly be already installed,and hence
will refresh all bundles depending on Guava.

Also, this is not necessary.

This patch is not directly tied to BUG-6956 but it's a result of it

Change-Id: If2fff1c6326fb8b53786d59e94c364cf0f3292c3
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>
(cherry picked from commit 622a8cf869f13ad75777e39612c9a26374132e3b)