Update patch set 1
Patch Set 1:

> I believe it is always these three projects which need an
 > occasional manual update. Do their git repo settings have something
 > in common?

Not sure but aaa and netconf only started doing this recently. We used to only have this issue with odlparent. It's probably worth looking into we think there are some leads.

What's interesting though is if we branch master to another branch and started tracking the different branch suddenly this all works. Which is why we never have to do this on stable/boron and stable/beryllium branches.

Patch-set: 1