Updated git submodules
Project: netconf master 77cb41a716accda78004ca34f738884f8d4e6e74

Close missing resources when netconf-topolgy goes down

When the connection goes down, we are closing the NetconfDeviceCommunicator
instance but we aren't closing the RemoteDeviceHandler. I believe this is an
oversight, so implementing the Autocloseable interface in NetconfConnectorDTO
that will close both the resources. The consumer of this DTO can then call #close()
to release the resources.

Change-Id: I852a5aa29a73f8454aa1886990dfce9c7a019025
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>