Bug 4219: Add karaf4 related parents to fix-relativepaths.sh
The parent were added in https://git.opendaylight.org/gerrit/47484
This is NOT needed for https://git.opendaylight.org/gerrit/47884
validate-autorelease job to pass (as it builds whole odlparent before validating)
but autorelease-release WOULD break without this if karaf4-parent
is used from non-odlparent project.

Not sure about odlparent-karaf4-empty, but the karaf3 line was already there,
so adding karaf4 version is the safer option.

Change-Id: Ice1cfe54f4e4e737448093aab6a2649f09a86c02
Signed-off-by: Vratko Polak <vrpolak@cisco.com>