Updated git submodules
Project: sfc master 69aa74ca82792841a4bf760e3f12d61539a2557e

Merge "Integrate classifier w/ genius - commit #3"

Integrate classifier w/ genius - commit #3

This commit finishes the integration of the current SFC classifier w/
genius.

The remaining methods of the API providing in the previous commits are
implemented, specifically:

The genius service binding w/ the interfaces to classify is finished,
and is called from the OpenflowClassifierProcessor.

The LogicalClassifierDataGetter is finished, providing all necessary
data-retrieval methods required by the LogicallyAttachedClassifier.

The flow generation logic made in the LogicallyAttachedClassifier is
finished.

Unit testing of the remaining classes is added.

Quicker bail-out of the openflowWriter when no flows are to be writen or
deleted.

Change-Id: I52755b98fc3eeb2f78c8afd90f4d83c6f65ae527
Signed-off-by: Miguel Duarte <miguel.duarte.de.mora.barroso@ericsson.com>