Updated git submodules
Project: yangtools master 815b544de729c2e7785431541236a1c679c222d7

Bug 6870: [Yang 1.1] Support for new "modifier" statement

The "modifier" statement, which is an optional substatement to the
"pattern" statement, takes as an argument the string "invert-match".
If a pattern has the "invert-match" modifier present, the type is
restricted to values that do not match the pattern.

Change-Id: I9e9d993a16077c1f3af1ab20c197d2e6c9e8c6af
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>