Updated git submodules
Project: sfc master 0e0851e7799ef9f52f0752c6562308330c024d01

Handle RSP updates on the classifier APIs

The APIs leveraged by the classifier were changed enabling it to update
its flows when the first SF in an RSP migrates to another compute node.

New classes (to be implemented in the next commit):
  - ClassifierRspUpdateDataGetter
This class queries MD-SAL for all ACLs whose ACEs target the supplied
RSPs, and also queries MD-SAL for all Classifiers enforcing the filtered
ACLs.

  - ClassifierRspsUpdateListener
Identifies relevant RSP updates (RSPs whose 1st SF neutron port changed
data-plane IDs), and processes that RSP update - deleting the old
impacted flows, and creating new ones.

API changes:
  - OpenflowClassifierProcessor
The OpenflowClassifierProcessor now *has* a BareClassifier, and a
LogicallyAttachedClassifier object, injected through the constructor.

Which classifier(Bare vs LogicallyAttached) is used for each processed
request depends on the type of SFF to which it is connected to.

  - SfcNshHeader
The SfcNshHeader class now encapsulates an RSP object, which can be set
through a different constructor. Using this allows us to save DS reads
on RSP updates, since we are notified with the new object.

  - SfcScfOfProcessor
The SfcScfOfProcessor class got simpler, and now does not have to
encapsulate and inject the transaction object into the OpenflowWriter
and OpenflowClassifierProcessor.

  - SfcScfOfRenderer
The SfcScfOfRenderer class now also features an update handler for RSPs.
Also, it injects the required dependencies into the
OpenflowClassifierProcessor and OpenflowWriter objects, while making
them share the same transaction objects.

  - SfcOfFlowWriterInterface
The SfcOfFlowWriterInterface and its subclass now encapsulate an
EventBus, and expose a transactional API. Clients that want to share the
same transaction object as the OpenflowWriter register themselves, and
are notified whenever the OpenflowWriter commits a transaction.

Also, the FlowDetails class manipulated by the SfcOfFlowWriterInterface
now encapsulates the used RSP path ID, making it easier / more efficient
for the SfcOfFlowWriterInterface clients to delete a set of related
flows.

Log messages are tuned throughout the classifier code, along with the
used log levels.

Change-Id: I0d61f090bdfec2e58aa0e811988ea9cca32be66c
Signed-off-by: Miguel Duarte <miguel.duarte.de.mora.barroso@ericsson.com>