Updated git submodules
Project: yangtools master 47704af447846a88bc7579a8151af247b7433f9b

BUG-7474: produce Karaf 4 features

This is a simplified version of
https://git.opendaylight.org/gerrit/38638 without any changes to
bundles; it just delegates the actual features.xml generation to the
plugin. This requires a current snapshot of Karaf 4.0.9 to reduce the
resulting dependencies. The new features are generated for Karaf 4
only. (The existing features.xml is preserved for Karaf 3.)

Change-Id: I4cc4692459f637e06e64a1a2e432448f57c32e4c
Signed-off-by: Stephen Kitt <skitt@redhat.com>