Updated git submodules
Project: yangtools master 91990bcb553eb75e6f242c8ad1c7e96afffd5da5

BUG-6522: cache isConfiguration() result

Profiling runs have identified the recursive search for
default config statements as a potential hotspot.

The problem here is that config statements are not common,
hence most of the time we end up traversing up the to the
root context (which is hard-coded config=true) and walk
the declared and effective statements over and over.

By the time this method is called its result should be
invariant, hence add a cache field and use it to side-step
both the recursion and substatement list walk.

Change-Id: I51383a9af0a7594b1d56b954ca9bbda0933b7231
Signed-off-by: Robert Varga <rovarga@cisco.com>