Updated git submodules
Project: openflowplugin master 5d7d942279274850a27d101e4b6c14ab6ad50c51

Bug 7415 Reducing the severity of the log message

Whenever there is a change in configuration of l2switch,
bundle refresh happens. When bundles refresh, openflowplugin
pushes the same flow. Changing the severity of of the exception
to warning since it is expected to have a flow with same id.

Change-Id: I69874d14b83053edc97b6d299c8fdb05d7ccdd8c
Signed-off-by: Sai MarapaReddy <sai.marapareddy@gmail.com>
Author: Sai MarapaReddy <sai.marapareddy@gmail.com>